[RFC] Make eslint do something
This commit enables eslint's indent check to produce more useful --fix
results, and also turns on the recommended set and fixes the resulting
(non-indent) complaints when it makes sense and disables them otherwise.

It is utterly pointless in that it does not change anything
in the program's behavior.

PS: the package-lock.json feels outdated, but I don't quite have the
energy to figure out what changed.

Close #105.