[types] Merge `Color` types
The types originally had two definitions for the `Color` type which
weren't interchangeable. This PR changes the types to more accurately
reflect the original JS by augmenting the existing `Color` type when
`src/index.d.ts` is imported.

There is a problem with how I've done this right now: the module
augmentation is happening even when the normal `src/color.d.ts` version
of the class is imported. I'm not sure how this is happening without
`src/index.d.ts` even being referenced, so I'll have to debug that when
I get the chance.