[types] Add `.js` extension to file imports (#333)
* [types] Add `.js` extension to file imports

Fixes problems when `moduleResolution` in tsconfig.json is set to
NodeNext.

* [types] Add extensions to `import()` types

* [types] Add missing imports

Weirdly, TypeScript only complains about these imports when
`moduleResolution` is set to `NodeNext`. I guess that whatever the
default module resolution system is can infer that they're from the
augmented module, but that doesn't seem to work on `NodeNext`.