Add gamutSpace to ColorSpace
Some colorspaces such as hsluv, okhsl and okhsv need to do
gamut checking in a color space that is not their own space or
their base space. For example hsluv needs to do gamut checking in
the srgb color space.

A gamutSpace can now be specified when constructing a ColorSpace and
a gamutSpace property has beed added to ColorSpace.

If a gamutSpace is not specified the following rules are used to
assign the gamutSpace property:

- if the color space is polar use the base space
- if the color space is not polar use the current space

Polar spaces that need to gamut check themselves (hpluv) can set the
gamutSpace option to "self" to avoid gamut checking using the base
space.