Allow hsluv to be gamut mapped using the CSS algorithm
The hsluv color space wasn't being gamut mapped by the CSS
algorithm because it was considered to be an unbounded space
and the CSS algorithm exits early if a space is unbounded.

Changing the 's' and 'l' coordinates to use range instead
of refRange makes hsluv a bounded space and allows it to be
gamut mapped properly.

Note that testing if a hsluv color was in gamut worked properly
because the gamutSpace of hsluv is srgb which is a bounded space.