Add new GMA Scale LH Achromatic
Performs a similar approach as Scale LH, but instead scales color
towards the achromatic version of itself. This is performed in a
linear light RGB space. When scaling the color towards the achromatic
color, it is done in an LRGB form that separates the lightness from
the other components. This helps preserve lightness during the process
and performs even better, especially in low light blue, which was more
problematic in our first attempt.

The procedure is competitive with Scale LH using two iterations,
performs even better with 4. 3 iterations was chosen as a middle ground.