Allow any color type when calling individual deltaE functions (#451)
* Allow any color type when calling individual deltaE functions

The types for the deltaE functions didn't allow all ColorTypes
as parameters. For example, the types didn't allow "red" to be
passed as a parameter.

Also, calling a deletaE method with a color such as "red"
would cause an exception to be thrown.

e.g. let x = deltaE2000("red", "white") // Exception thrown

* Pass arrays of colors to getColor()