Final clip is required in Ray tracing
This is probably as far as we can take ray tracing.

It was though that the last correction and clip could be omitted, but
it is required to get the optimal deltas. This time confirmed with
a script. Max deltas: ∆L = 0.011682128741787201 ∆h = -4.212684767726387

Previously, there was still a region (in yellow) that could exceed this.
(before ∆L = 0.0011833203080477972 ∆h = -8.31632698148956).