Fix perpendicular corner case in ray trace algorithm
The found algorithm suffered from a case where the algorithm breaks down
with a perfectly perpendicular case. There are a few other related cases
as well. When these cases are identified, ensure they are identified,
and passed through properly, also calculate the intersection with the
cube via inverse interpolation.

This seem less likely to occur, but should be handled properly if/when
they do occur.