Simplify logic based on how SLAB works
SLAB will find the near and the far intersection past the end points
of the ray. We were clamping it prior. Allow the ray to extend, but
ensure the intersection that is returned is the first in the direction
start -> end. This allows us to simplify the code a bit more as it can
automatically search the intersection towards and away from the
achromatic point based on the origin without us doing extra work.