Optimize to be a little faster and a little more accurate
This is likely the last tweak as we've probably optimized this as much
as we can.

Optimize code to generally be faster so that we can improve accuracy
without a net decrease in speed.

- Remove unnecessary clone calls and unused variables
- Slab allows us to always ray trace from the inside of the cube, so
  simplify code and always do that, results are identical as inside
  out gives us the best results as we always want to find the cube side
  in the direction of the corrected color looking out from the
  achromatic origin.
- Corrections to L and H are done with an oklch proxy which is faster
  and valid as long as we don't update the original color directly.
- Avoid loop when clamping.

Generally, performance is slightly faster with an overall improvement to
deltas. Very slight increase of lightness delta with a significant
decrease of hue delta.

∆L = -0.0007490954865150279
∆h = 2.5244051656294317