Faster ray trace GMA
Ensure the initial ray cast is done with the smallest angle. This is
done by passing the origin (color with no chroma) and passing the ray
through a second color with (almost no chroma) to guide the direction.
This is done instead of casting through the origin and the current color
which often yields the greatest angle of deviation. This will actually
cause the GMA to converge on an optimal value in less iterations,
allowing us to only require 3 passes instead of 4 while still
maintaining a ∆h somewhere in the ~2 range.