[types] overload multiplyMatrices types (#580)
* [types] overload multiplyMatrices types

Adds overloads for the parameter and return types of multiplyMatrices.
And tests in `types/test` for the overloads.

Includes updates to `adapt.js` and `CATs.js` to fix type errors due to
the new overloaded types.

I was probably a bit too overzealous and added functionality for
returning a number result as well. Let me know if this should be in a
separate pull request.

* [types] remove scalar value return functionality

Functionality changes will occur in separate commits.