Set typescript target of ES2018
Typescript 5.5 add support for regular expression syntax checking.
One of the checks tests if a regular expression is using features
that aren't supported by the target version of ECMAScript.

Named capture groups are used in src/Type.js and possibly some
other files so a target of ES2018 was added to tsconfig.json.