Avoid style recalculation on scroll events
The previous code depended on getting the scrolled amount through a custom property on the root element.

This caused laggy scrolling and can be avoided by using `position: sticky`, which has the same behavior as the previous code.

The best way to do this is by changing the markup to be similar to the `/api` pages, which use a separate stylesheet. It uses a grid that allows the sticky element to have space next to the main content. For the other pages, the submenu is actually inside of the main content container and relies on `position: fixed` to escape it.

However such a markup change would need a lot of CSS adaptations. Hence this commit solves it with CSS only instead, in a way that is very unlikely to affect other elements on the page.