Avoid full style recalculation by setting custom property only on element that uses it
Since the `--scrolltop` custom property is only used on the `#toc` element,
we can avoid a *full* style recalculation of all elements on the page which can take very long,
and causes massive framerate drops during scrolling.

This should not result in any changed behavior or appearance as no other element uses the custom property.