Replace `instanceof` checks to work across color.js sources
Instead of relying solely on `instanceof` to check if an argument is an
instance of some class (eg. Color or ColorSpace), also checks the name
of the constructors of the instance to see if any of them have the same
name as the target class.

This method isn't perfect, but it should work well enough given that it
seems unlikely for a user of color.js to define their own unrelated
class whose name conflicts with the library.