Fix jzczhz check for achromatic colors
The default ε (0.00001) calculated for jzczhz is to low for
achromatic colors. The highest az or bz value for an achromatic
color is -0.00023625807422922307 for color(rec2100-pq 1 1 1).

The ColorSpace class has been modified to allow setting ε when
constructing a ColorSpace and the ε for jzczhz has been set to
0.0002363.

This new ε ensures that all achromatic colors up to the brightest
white in req2100-pq have a null hue when converted to jzczhz.

However, with the new ε there are some non-achromatic colors near
black in rec2100-pq that will be considered to be achromatic when
converted to jzczhz.