fix: parser not being able to distinguish between legacy and modern formats
Validate the parsed CSS types for elements in color strings that are using functional syntax against the coordinate grammar defined in the corresponding color space. If a disallowed type is found in the parsed color, a `TypeError` is thrown.

To actually distinguish legacy and modern syntaxes, the parser was updated to look for a named format with a `_legacy` suffix first (e.g. parsing an `hsl` color will look for `hsl_legacy` first and then for `hsl`). This is not a good system as it's not flexible and because it relies on somewhat strangely named formats to be defined in the library. Ideally, color spaces define their grammar kind of like the specification does. The parser would identify the color space and then check if the parsed color matches any of the space's grammar(s) (e.g. for HSL one of legacy hsl/hsla or modern hsl/hsla). I also think it would be useful to distinguish between grammars (parsing) and formats (serialization). However, any further structural changes to the parsing logic would blow up the scope of this change which intends to fix the incorrect parsing of legacy HSL colors.

Closes #648.