Fix return types of the darken/lighten functions on the Color class (#654)
Changes the types of the lighten and darken instance and static
functions to return Color instead of PlainColorObject.

Also fixes an issue with the static functions returning a
PlainColorObject instead of a Color when anything but
a Color object was passed to them. For example:

new Color("red").lighten() would return a PlainColorObject
instead of a Color object.