[types] Use `@import` instead of `@typedef` (#686)
* Bump TypeScript to v5.5

To allow us to use the `@import` keyword. Also bumps TypeDoc to a
version that supports TypeScript v5.5.

* [types] Replace `@typedef` with `@import` where relevant

* [types] Change types that should be re-exported back to `@typedef`

As of now, the latest TypeScript version has no export equivalent of
`@import`, so the only way to expose a type from another file seems to
be through `@typedef`s.

* Style fixes

Mostly adding/removing whitespace that Prettier didn't change.

* [types] Replace `default as` type imports

Semantically the same, but matches the typical way of importing ESM
defaults.

* [types] Address review comments

Co-Authored-By: Lloyd Kupchanko <lloyd@lloydk.ca>

---------

Co-authored-by: Lloyd Kupchanko <lloyd@lloydk.ca>