[ci] Use `npm ci` instead of `npm install`
Prevents npm from touching the lockfile when running in CI, which fixes
the CI error we've been having.

Also adds a call to `git diff` in case of errors so that it's easier to
spot the reason for CI failures in the future.

See: https://docs.npmjs.com/cli/v11/commands/npm-ci